ai_api/ai_api/requirements.txt
flask==2.3.0
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.0
requests==2.31.0
